context available in initState
Yes, you can access a provider, like AppState, in the initState method of a State class, but there is a small caveat. In Flutter, you should avoid accessing the BuildContext directly during the initial setup of the widget, because it may not be fully initialized when initState is called.
However, when you're using the Provider package, it's designed to be safe and ensure that accessing the BuildContext is fine as long as the widget has been added to the widget tree, which is generally the case when initState is being called. This is why context.read<AppState>() works without issues.